Moving InventoryOperationResult to its own file
