+++ Assets/Scripts/Inventory/InventoryOperationResult.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 23f06a8b93427224b858eb7d558b1869
